+++ source/http_request_builder.h
@@ -58,7 +58,7 @@ public:
         }
     }
 
-    char* build(const void* body, size_t body_size, size_t &size, bool skip_content_length = false) {
+    char* build(const void* body, unsigned int body_size, uint32_t &size, bool skip_content_length = false) {
         const char* method_str = http_method_str(method);
 
         bool is_chunked = has_header("Transfer-Encoding", "chunked");
